Fixed compatibility with ISO C++. Expanded function parameter trace to offer multiple levels of verbosity.

+++ src/bigrfc.h
@@ -58,9 +58,9 @@ SEXP treepredict(BigMatrix* x, int ntest, SEXP forestP, SEXP treeP);
     
 /* Utility function for "unpacking"" an integer into an array of 1s and 0s (i.e.
    binary). */
-void unpack(unsigned long long npack, int icat[], int l);
+void unpack(unsigned long npack, int icat[], int l);
 
 /* The reverse of unpack. */
-unsigned long long pack(int icat[], int l);
+unsigned long pack(int icat[], int l);
 
 #endif
